pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voiceclone"
version = "0.2.0"
description = "Voice cloning and TTS narration on Apple Silicon (Qwen3-TTS + MOSS-TTS)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Gregory Gorin" }]
dependencies = [
  # Primary: Qwen3-TTS via mlx-audio
  "mlx-audio",
  "soundfile",
  # Post-processing
  "pedalboard",
  # Shared
  "numpy",
  "scipy",
  "pyyaml",
  "tqdm",
  # Legacy: MOSS-TTS via transformers
  "torch>=2.5.0",
  "torchaudio>=2.5.0",
  "transformers>=5.0.0",
  "safetensors",
  "orjson",
  "einops",
  "librosa",
  "tiktoken",
  "psutil",
  "packaging",
]

[project.optional-dependencies]
flash-attn = ["flash-attn"]
gradio = ["gradio"]
whisper = ["openai-whisper"]

[tool.setuptools]
py-modules = []
